Update hyprland.conf
leaving vfr enabled

It has nothing to do with damage tracking.

VFR: New frame is submitted to the pipeline when there is a need for one (something changed and we need to redraw)
No VFR: New frames are submitted every time your monitor has a vblank. This does not mean we redraw everything, if there's nothing to redraw, we'll submit the last drawn frame. This has a serious effect on the CPU usage, though, as the "submitting" takes a bit of CPU.